refactor(command-bar): make vmux_command the page it already served
vmux_command owned COMMAND_BAR_PAGE_URL, CommandBarActionEvent,
CommandBarOpenEvent, the snapshots and the contributions registry — every
part of vmux://command-bar/ except the two files that render it, which sat
in vmux_layout because that is where the command bar grew up.

Those two move here and the crate moves to page/, where AGENTS.md says a
crate that answers a URL belongs. What stays in vmux_layout is handler.rs:
deciding what a chosen row means is stack and pane work, and moving it
would invert the dependency every other crate already has on this one.

Less had to move than the line count suggested. keyboard, results and style
were already vmux_start's, re-exported through command_bar.rs, so the page
half was page.rs and palette.rs alone.

Three things the palette reached for had to move out of vmux_layout first,
each a wire type or a pure function that was only there because the palette
was: StartSelectWorkspace and supports_inline_agent_transition to vmux_wire,
BookmarksCommandEvent to this crate's event module — it needs PageMetadata,
which vmux_wire cannot see, since vmux_core depends on vmux_wire and not the
other way round. All three are re-exported from where they were, so nothing
downstream changed.

size.rs came along and stays ungated. It was unconditional in vmux_layout
and gating it behind cfg(web) here quietly stopped its two tests from
running at all. Its 7-tuple, passed to should_emit and then to mark_emitted
in the same order, becomes CommandBarSize — clippy asked for one fewer
argument and the type was the missing thing, not the lint.
