+++ crates/host/vmux_command_mcp/Cargo.toml
@@ -10,4 +10,4 @@ serde_json = { workspace = true }
 vmux_macro = { path = "../../vmux_macro" }
 
 [dev-dependencies]
-vmux_command = { path = "../../vmux_command" }
+vmux_command = { path = "../../page/vmux_command" }
