+++ crates/page/vmux_command/assets/index.html
