+++ crates/page/vmux_command/src/host.rs
@@ -5,6 +5,20 @@
 //! unchanged — `lib.rs` re-exports this module's contents, so `vmux_command::open` still resolves
 //! from outside and `crate::open` still resolves from within.
 
+/// The `vmux://command-bar/` page this crate serves.
+///
+/// `command_bar: false` because the launcher does not list itself.
+pub const COMMAND_BAR_PAGE_MANIFEST: vmux_core::page::PageManifest =
+    vmux_core::page::PageManifest {
+        host: "command-bar",
+        title: "Command Bar",
+        title_message_id: None,
+        replaces_command: None,
+        keywords: &[],
+        icon: None,
+        command_bar: false,
+    };
+
 pub mod plugin;
 pub use plugin::CommandPlugin;
