+++ crates/page/vmux_command/tests/open_command_derives.rs
