+++ crates/page/vmux_layout/Cargo.toml
@@ -39,7 +39,7 @@ rfd = { workspace = true }
 ron = { workspace = true }
 sha2 = "0.10"
 uuid = { workspace = true }
-vmux_command = { path = "../../vmux_command" }
+vmux_command = { path = "../vmux_command" }
 vmux_git = { path = "../../vmux_git" }
 vmux_history = { path = "../vmux_history" }
 
@@ -57,7 +57,7 @@ objc2-app-kit = { version = "0.3", default-features = false, features = [
 dioxus = { workspace = true }
 qrcode = { version = "0.14", default-features = false, features = ["svg"] }
 serde_json = { workspace = true }
-vmux_command = { path = "../../vmux_command" }
+vmux_command = { path = "../vmux_command" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 gloo-timers = { version = "0.3", features = ["futures"] }
