+++ crates/page/vmux_layout/src/command_bar.rs
@@ -1,12 +1,3 @@
-#[cfg(web)]
-pub mod page;
-#[cfg(web)]
-pub mod palette;
-
-pub use vmux_start::{keyboard, results, style};
-
-pub mod size;
-
 #[cfg(host)]
 pub mod handler;
 #[cfg(host)]
