+++ crates/page/vmux_team/Cargo.toml
@@ -26,7 +26,7 @@ bevy_cef = { workspace = true }
 bevy_ecs = { workspace = true }
 ron = { workspace = true }
 vmux_agent = { path = "../vmux_agent" }
-vmux_command = { path = "../../vmux_command" }
+vmux_command = { path = "../vmux_command" }
 vmux_layout = { path = "../vmux_layout" }
 vmux_space = { path = "../vmux_space" }
