+++ crates/page/vmux_terminal/Cargo.toml
@@ -27,7 +27,7 @@ rfd = { workspace = true }
 sysinfo = "0.38"
 tracing = "0.1"
 url = { workspace = true }
-vmux_command = { path = "../../vmux_command" }
+vmux_command = { path = "../vmux_command" }
 vmux_history = { path = "../vmux_history" }
 vmux_layout = { path = "../vmux_layout" }
 vmux_service = { path = "../../host/vmux_service" }
