--- crates/vmux_command/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "vmux_command"
-description = "Command vocab + command-bar wire events"
-version.workspace = true
-edition.workspace = true
-publish = false
-build = "../build_cfg_git.rs"
-
-[features]
-default = []
-debug = []
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-rkyv = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-url = { workspace = true }
-vmux_core = { path = "../vmux_core" }
-vmux_history = { path = "../page/vmux_history", default-features = false }
-vmux_wire = { path = "../vmux_wire" }
-
-[target.'cfg(not(any(target_arch = "wasm32", target_os = "ios")))'.dependencies]
-bevy = { workspace = true }
-bevy_cef = { workspace = true }
-bevy_ecs = { workspace = true }
-muda = "0.19"
-vmux_macro = { path = "../vmux_macro" }
