+++ crates/vmux_page/Cargo.toml
@@ -15,6 +15,7 @@ web = [
     "dep:vmux_chat",
     "dep:vmux_editor",
     "dep:vmux_history",
+    "dep:vmux_command",
     "dep:vmux_layout",
     "dep:vmux_service",
     "dep:vmux_setting",
@@ -45,6 +46,7 @@ vmux_agent = { path = "../page/vmux_agent", optional = true }
 vmux_chat = { path = "../page/vmux_chat", optional = true }
 vmux_editor = { path = "../page/vmux_editor", optional = true }
 vmux_history = { path = "../page/vmux_history", optional = true }
+vmux_command = { path = "../page/vmux_command", optional = true }
 vmux_layout = { path = "../page/vmux_layout", optional = true }
 vmux_service = { path = "../host/vmux_service", optional = true }
 vmux_setting = { path = "../page/vmux_setting", optional = true }
