+++ crates/vmux_page/src/web.rs
@@ -56,7 +56,7 @@ macro_rules! web_pages {
 
 web_pages! {
     render_layout: "layout" => vmux_layout::page::Page,
-    render_command_bar: "command-bar" => vmux_layout::command_bar::page::Page,
+    render_command_bar: "command-bar" => vmux_command::page::Page,
     render_debug: "debug" => vmux_layout::debug_page::Page,
     render_error: "error" => vmux_layout::error_page::Page,
     render_terminal: "terminal" => vmux_terminal::page::Page,
@@ -107,7 +107,7 @@ fn StartAgentPage() -> Element {
     }
     rsx! {
         vmux_layout::start::page::Page {
-            on_inline_transition: move |next: vmux_layout::command_bar::palette::StartInlineTransition| {
+            on_inline_transition: move |next: vmux_command::palette::StartInlineTransition| {
                 vmux_layout::start::page::begin_agent_transition();
                 InlineAgentWindow::set(&next.target_url);
                 transition.set(Some(next));
@@ -124,10 +124,10 @@ impl InlineAgentWindow {
     const PREFIX: &'static str = "vmux-inline-agent:";
 
     /// The transition this window was opened for, if it was opened for one.
-    fn pending() -> Option<vmux_layout::command_bar::palette::StartInlineTransition> {
+    fn pending() -> Option<vmux_command::palette::StartInlineTransition> {
         let name = web_sys::window()?.name().ok()?;
         let agent_url = name.strip_prefix(Self::PREFIX)?;
-        Some(vmux_layout::command_bar::palette::StartInlineTransition {
+        Some(vmux_command::palette::StartInlineTransition {
             target_url: agent_url.to_string(),
             prompt: String::new(),
             attachments: Vec::new(),
