+++ crates/vmux_wire/src/agent.rs
@@ -71,6 +71,16 @@ impl AgentKind {
     }
 }
 
+/// Whether an agent page can replace the launcher inside its existing webview.
+pub fn supports_inline_agent_transition(url: &str) -> bool {
+    let Some(path) = url.strip_prefix("vmux://agent/") else {
+        return false;
+    };
+    !path
+        .split('/')
+        .any(|segment| matches!(segment, "cli" | "setup"))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
